requirements.txt
streamlit
cv2
json
dotenv
moviepy.video.io.ffmpeg_tools
moviepy.editor
openai
base64
mimetypes
yt_dlp
yt_dlp.utils